third_party/WebKit/LayoutTests/svg/dom/SVGLengthList-getItem-expected.txt

ABC
This is a test of the SVGLengthList::getItem() API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check lengths of text x item lists
PASS text1.x.baseVal.numberOfItems is 3

Test with various index parameters
PASS text1.x.baseVal.getItem(-100) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGLengthList': The index provided (4294967196) is greater than the maximum bound (3)..
PASS text1.x.baseVal.getItem(-1) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGLengthList': The index provided (4294967295) is greater than the maximum bound (3)..
PASS text1.x.baseVal.getItem(0).value is 50
PASS text1.x.baseVal.getItem(1).value is 100
PASS text1.x.baseVal.getItem(2).value is 150
PASS text1.x.baseVal.getItem(3) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGLengthList': The index provided (3) is greater than or equal to the maximum bound (3)..
PASS text1.x.baseVal.getItem(100) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGLengthList': The index provided (100) is greater than the maximum bound (3)..
PASS successfullyParsed is true

TEST COMPLETE
